convert datadogclient to component
