+++ comp/core/datadogclient/component.go
@@ -1,0 +1,19 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+package datadogclient
+
+import (
+	"gopkg.in/zorkian/go-datadog-api.v2"
+)
+
+// Component is the component type.
+type Component interface {
+	// QueryMetrics takes as input from, to (seconds from Unix Epoch) and query string and then requests
+	// timeseries data for that time peried
+	QueryMetrics(from, to int64, query string) ([]datadog.Series, error)
+
+	// GetRateLimitStats is a threadsafe getter to retrieve the rate limiting stats associated with the Client.
+	GetRateLimitStats() map[string]datadog.RateLimit
+}
